Добавление модуля генерации объявлений
